Monkey-patch list/delete race
GeeseFS may issue LIST and DELETE requests at the same time and without
this fix it may be possible that DELETE is executed after LIST by the server,
but DELETE response is delivered first. In this case an inode may become
deleted and then "resurrected" by the LIST response handler.

It leads to xfstests sometimes failing because they try to move the object
that, in fact, no longer exists. Maybe it could be cool to solve this problem
by using something clever (replicated log?), but for now I use a simple fix:
the patch logs all deletions in a separate list when some listings are in
progress and applies them to all listing results.