Restore directory position using name after changes
This fixes possible duplicates in listings caused by the removal
of "stale" inodes when the listing starts

Again, it'd be better to not remove and re-add all files on each
reload of a directory, i.e. remove only files that don't exist
on server anymore, but this fix is being prepared in the `rm-only-changed`
branch and it's not done yet. And restoring directory position using name
is better anyway :-)